requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-braces==1.16.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
oauthlib==3.2.2
pillow==10.3.0
PyJWT==2.9.0
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
